Wait for the creation of the cluster to continue the operations
Creating a cluster can take a few minutes. The "create" operation now
waits for the cluster to be "healthy" before continuing.

The Terraform plugin framework doesn't contain any support for retries
and timeouts, so this is all a bit hardcoded for now.

See: https://github.com/hashicorp/terraform-plugin-framework/issues/400

The code was inspired from the Terraform SDK documentation:
https://www.terraform.io/plugin/sdkv2/resources/retries-and-customizable-timeouts#statechangeconf